Move example stories into `example/` directory
This was a miss from 50679e689491e39cb6090b86275dd3198219cb6d because I
didn’t understand the importance of story placement, too. Apologies!

Nest example stories within an `example/` directory as well, both for
semantic reasons and so we can write stories about our own components
with conflicting names in the top-level namespace.

Issue #10 Scaffold routing and page construction
