+++ src/stories/example/Page.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { within, userEvent } from '@storybook/testing-library';
 
-import { Page } from '../pages/example/Page';
+import { Page } from '../../pages/example/Page';
 
 const meta = {
   title: 'Example/Page',
